+++ src/main/java/cart/service/AdminService.java
@@ -3,7 +3,6 @@ package cart.service;
 import cart.dao.ProductDao;
 import cart.dto.ProductRequestDto;
 import cart.entity.ProductEntity;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
